requirements.txt
farm-haystack
sentence-transformers
torch 